services/auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyjwt
passlib[argon2]
python-multipart
email-validator
py-eureka-client
